Revert "separate into different tests"
This reverts commit 8e3eb7a00356a874415e0071dd9968334cdd9a75.
