+++ tests/functional/plugins/test_motherduck_attach.py
@@ -104,3 +104,52 @@ class TestMDPluginAttach:
 
     def test_motherduck(self, project):
         run_dbt(expect_pass=False)
+
+
+@pytest.mark.skip_profile("buenavista", "file", "memory")
+class TestMDPluginAttachWithSettings(TestMDPluginAttach):
+    @pytest.fixture(scope="class")
+    def profiles_config_update(self, dbt_profile_target):
+        md_setting = {"motherduck_token": dbt_profile_target.get("token")}
+        return {
+            "test": {
+                "outputs": {
+                    "dev": {
+                        "type": "duckdb",
+                        "path": ":memory:",
+                        "attach": [
+                            {
+                                "path": dbt_profile_target.get("path", ":memory:"),
+                                "type": "motherduck"
+                            }
+                        ],
+                        "settings": md_setting
+                    }
+                },
+                "target": "dev",
+            }
+        }
+
+
+@pytest.mark.skip_profile("buenavista", "file", "memory")
+class TestMDPluginAttachWithTokenInPath(TestMDPluginAttach):
+    @pytest.fixture(scope="class")
+    def profiles_config_update(self, dbt_profile_target):
+        token = dbt_profile_target.get("token")
+        return {
+            "test": {
+                "outputs": {
+                    "dev": {
+                        "type": "duckdb",
+                        "path": ":memory:",
+                        "attach": [
+                            {
+                                "path": dbt_profile_target.get("path", ":memory:") + f"?motherduck_token={token}&user=1",
+                                "type": "motherduck"
+                            }
+                        ]
+                    }
+                },
+                "target": "dev",
+            }
+        }
